revisao estrutura dos conteudos das semanas

+++ semana-2/src/calculator3.c
@@ -22,7 +22,6 @@ int main(void)
   printf("x: "); scanf("%d", &x);
   printf("y: "); scanf("%d", &y);
 
-  float z = x / y; 
-  // float z = (float) x / (float) y;
+  float z = (int) x / (float) y;
   printf("%f\n", z);
 }
